+++ components/deprecated/index.js
@@ -1,0 +1,1 @@
+export * as TextInput from './text-input';