+++ webpack.config.js
@@ -5,6 +5,7 @@ const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 module.exports = {
 	entry: {
 		main: './components/main.js',
+		deprecated: './components/deprecated/index.js',
 		'ag-grid': './components/grid/index.js',
 	},
 	devtool: 'sourcemap',