+++ VersionHistory.md
@@ -4,6 +4,8 @@
 
 Please add changes here when they're committed to the `master` branch, then move to "Released" once the package is published.
 
+* Moved TextInput components to deprecated bundle. They will be dropped on the next major release. Nobody was using these, but it's technically a breaking change. If you really need them back, import them from '@faithlife/styled-ui/dist/deprecated.js'
+
 ## Released
 
 ### 1.0.0